+++ gzap/apworld/gzdoom/Options.py
@@ -96,10 +96,11 @@ class StartingLevels(OptionSet):
     """
     Set of levels to begin with access to. If you select levels that aren't in the
     WAD you choose (e.g. E1M1 when you're generating for Doom 2) they will be safely
-    ignored. If you don't select any levels in your chosen WAD, it will force you
-    to start with access to the first level.
+    ignored. You must select at least one level in your chosen WAD or generation
+    will fail.
 
-    You will begin with the access code and all keys for these levels in your inventory.
+    You will start with the access codes for these levels; if start_with_keys is
+    enabled, you will also start with all keys for them.
 
     This option supports globbing expressions.
     """